move logger to the gui_tofCams_bridge
